Split the Anaconda configuration handler to more files
The file that defines the Anaconda configuration handler should be
split to several smaller files, because it started to be too large.
